requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.3
frozenlist==1.4.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
numpy==2.0.0
openpyxl==3.1.4
pandas==2.2.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
twilio==9.2.1
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
Flask==2.0.1
gunicorn==20.1.0
pandas==1.3.3
openpyxl==3.0.7
twilio==6.63.0
pdfplumber==0.5.28
Pillow==8.3.2
openai==0.10.1
